pyproject.toml
[tool.poetry]
name = "recommendation-system-ps"
version = "0.1.0"
description = ""
authors = ["Ehor Hleoza <gleoza04@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
scikit-surprise = "^1.1.4"
numpy = "1.24"
scipy = "^1.14.1"
py-sc-kpm = "^0.4.0"
py-sc-client = "^0.4.0"
staticmap = "^0.5.7"
geopy = "^2.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
